moved toolbar components to own file, made dl yaml seperate component
